# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": ""
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/positions": {
            "get": {
                "tags": [
                    "POSITIONS"
                ],
                "summary": "Get all positions",
                "description": "Get all positions",
                "operationId": "getAllPositions",
                "responses": {
                    "200": {
                        "description": "Show all positions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "strind",
                                            "example": "true"
                                        },
                                        "positions": {
                                            "properties": {
                                                "id": {
                                                    "description": "Position ID",
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "description": "Position name",
                                                    "type": "string",
                                                    "example": "Developer"
                                                }
                                            },
                                            "type": "object",
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Positions not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Positions not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/token": {
            "get": {
                "tags": [
                    "TOKEN"
                ],
                "summary": "Get token",
                "description": "Get token",
                "operationId": "getToken",
                "responses": {
                    "200": {
                        "description": "Bearer Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Operation status",
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "token": {
                                            "description": "Token",
                                            "type": "string",
                                            "example": "9|VGRydIOV9tmF0BPwDAl1FHolXLsh9CmuvvmP2rYb"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "USERS"
                ],
                "summary": "Get all users",
                "description": "Get all users",
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Specify the page that you want to retrieve",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Specify the missing record number",
                        "required": false,
                        "example": "0"
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Specify the amount of items that will be retrieved per page",
                        "required": false,
                        "example": "5"
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "USERS"
                ],
                "summary": "Add new user",
                "description": "Add new user",
                "operationId": "addNewUser",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "name": {
                                        "description": "User's Name",
                                        "type": "string",
                                        "example": "Vasya"
                                    },
                                    "phone": {
                                        "description": "User's phone",
                                        "type": "string",
                                        "example": "+380671234567"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "example": "vasiliev@gmail.com"
                                    },
                                    "position_id": {
                                        "description": "Position ID",
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Operation status",
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "user_id": {
                                            "description": "User's id",
                                            "type": "integer",
                                            "example": "46"
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "New user successfully registered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "USERS"
                ],
                "summary": "Get user info by ID",
                "description": "Get user info by ID",
                "operationId": "getUserInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}